Added report about trace lifetimes
Not fully finished since I currently can only run it on a small portion
of the data (approx 250 traces) so that it cannot decisively rule out
distributions. This is because my trace plugins don't appear outside
debug mode and my own xes-importer is not disk-buffered so it cannot
handle the full dataset.